requirements.txt
Django>=5.0,<6.0
django-ninja>=1.1,<2.0
python-dotenv>=1.0,<2.0
firebase-admin>=6.5,<7.0
requests>=2.31,<3.0
gunicorn>=22.0,<23.0
uvicorn>=0.30,<1.0
psycopg[binary]>=3.1,<4.0
pytest>=8.2,<9.0
pytest-django>=4.8,<5.0
boto3>=1.34,<2.0
psycopg2-binary>=2.9,<3.0
